Added childs link in menu. Added child parent component; Already add parent to table after do some validations
